+++ superphy/src/upload/python/turtle_utils.py
@@ -1,0 +1,53 @@
+def generate_hash(filename):
+    from hashlib import sha1
+    # the 'b' isn't needed less you run this on Windows
+    with open(filename, 'rb') as f:
+        # we apply a sort func to make sure the contents are the same,
+        # regardless of order
+        return sha1(str(sorted(f.readlines()))).hexdigest()
+
+
+def generate_uri(uri, s=''):
+    """
+    Takes a string as one would define for .ttl files and returns a URI for rdflib.
+
+    Args:
+        uri (str): a string following .ttl convention for a URI
+        ex. g:Identifier as shorthand for http://www.biointerchange.org/gfvo#Identifier
+    Returns:
+        (rdflib.URIRef) with URI needed to add to rdflib.Graph
+    """
+    import settings  # this is the settings.py
+
+    from rdflib import Namespace, URIRef, Literal
+
+    # if you call with a uri already
+    if isinstance(uri, URIRef):
+        return URIRef(str(uri) + s)
+
+    prefix = uri.split(':')[0]
+    postfix = uri.split(':')[1]
+
+    if prefix == '':  # this is our : case
+        return URIRef(settings.namespaces['root'] + postfix)
+    else:
+        return URIRef(settings.namespaces[prefix] + postfix)
+
+
+def uri_to_basename(uri):
+    '''
+    This does the reverse of generate_uri(). Converts a rdflib.term.URIRef back to is base.
+        ex. rdflib.term.URIRef(u'https://www.github.com/superphy#4eb02f5676bc808f86c0f014bbce15775adf06ba)
+                gives 4eb02f5676bc808f86c0f014bbce15775adf06ba
+    Args:
+        uri(rdflib.term.URIRef): a URIRef object
+    Returns:
+        (str): just the basestring (ie. everything after the : in rdf syntax)
+    '''
+    import settings
+    for value in settings.namespaces.keys():
+        if value in uri:
+            return str(uri).strip(value)
+    # if the clean method above fails, default to '/' splitting
+    # this will fail if a path-style uri is used
+    return str(uri).split('/')[-1]
